move the env to the correct place
